feat: rename modules
